(DO NOT MERGE) Improving debug output and hardening 'send' contracts
The expected and actual behavior of the ros communication classes is being
hidden by a combination of swallowed stack traces and a function with the same
signature being used to invoke underlying functionality with different
contracts.

Added full stack trace print statements in cases where we were previously just
printing the name of the error to console and carrying on. Added a helper
function to better clarify what the expected behavior of `send` is for a given
class.

Tested against Pick and Place's demo scene -- NOT WORKING YET